+++ src/diskover/Diskover.php
@@ -12,7 +12,7 @@ use Elasticsearch\ClientBuilder;
 error_reporting(E_ALL ^ E_NOTICE);
 
 // diskover-web version
-$VERSION = '1.5.0.7';
+$VERSION = '1.5.0.8';
 
 
 function connectES() {